+++ src/input.cc
@@ -33,7 +33,6 @@ namespace redsea {
  *
  */
 MPXReader::MPXReader(const Options& options) :
-    input_type_(options.input_type),
     feed_thru_(options.feed_thru),
     sfinfo_({0, 0, 0, 0, 0, 0}) {
   is_eof_ = false;
@@ -42,7 +41,7 @@ MPXReader::MPXReader(const Options& options) :
       options.input_type == INPUT_MPX_SNDFILE) {
 
     if (options.input_type == INPUT_MPX_STDIN) {
-      sfinfo_.channels = 1;
+      sfinfo_.channels = options.num_channels;
       sfinfo_.format = SF_FORMAT_RAW | SF_FORMAT_PCM_16;
       sfinfo_.samplerate = options.samplerate;
       sfinfo_.frames = 0;
@@ -75,22 +74,33 @@ bool MPXReader::eof() const {
   return is_eof_;
 }
 
-std::vector<float> MPXReader::ReadChunk() {
+/*
+ * Fill the internal buffer with fresh samples. This should be called before
+ * the first channel is processed via ReadChunk().
+ *
+ */
+void MPXReader::FillBuffer() {
+  num_read_ = sf_read_float(file_, buffer_.data(), used_buffer_size_);
+  if (num_read_ < static_cast<long long>(used_buffer_size_))
+    is_eof_ = true;
+
+  if (feed_thru_)
+    sf_write_float(outfile_, buffer_.data(), num_read_);
+}
+
+std::vector<float> MPXReader::ReadChunk(int channel) {
+  assert(channel < sfinfo_.channels);
+
   std::vector<float> chunk;
   if (is_eof_)
     return chunk;
 
-  sf_count_t num_read =
-      sf_read_float(file_, buffer_.data(), used_buffer_size_);
-  if (num_read < static_cast<long long>(used_buffer_size_))
-    is_eof_ = true;
-
   if (sfinfo_.channels == 1) {
     chunk = std::vector<float>(buffer_.begin(), buffer_.end());
   } else {
-    chunk = std::vector<float>(num_read / sfinfo_.channels);
+    chunk = std::vector<float>(num_read_ / sfinfo_.channels);
     for (size_t i = 0; i < chunk.size(); i++)
-      chunk[i] = buffer_[i * sfinfo_.channels];
+      chunk[i] = buffer_[i * sfinfo_.channels + channel];
   }
   return chunk;
 }
@@ -99,6 +109,15 @@ float MPXReader::samplerate() const {
   return sfinfo_.samplerate;
 }
 
+int MPXReader::num_channels() const {
+  return sfinfo_.channels;
+}
+
+/*
+ * An AsciiBitReader reads an unsynchronized serial bitstream as '0' and '1'
+ * characters via stdin.
+ *
+ */
 AsciiBitReader::AsciiBitReader(const Options& options) :
     is_eof_(false), feed_thru_(options.feed_thru) {
 }